requirements.txt
langchain_community==0.2.9
fastapi==0.110.3
BeautifulSoup4==4.12.2
pydantic==2.8.2
langchain==0.2.10
langchain_groq==0.1.6
langchain_google_genai==1.0.7
python-dotenv==1.0.1
langchain_core==0.2.22